+++ src/core/services/geminiService.ts
