+++ src/core/services/userService.ts
@@ -1,8 +1,8 @@
 import { eq, sql } from "drizzle-orm";
 import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import type { User } from "@/types";
-import type * as schema from "../database/schema";
-import { users } from "../database/schema";
+import type * as schema from "../../database/schema";
+import { users } from "../../database/schema";
 
 export class UserService {
 	constructor(private database: BunSQLiteDatabase<typeof schema>) {}
@@ -39,12 +39,10 @@ export class UserService {
 	};
 
 	addLevel = async (userId: string): Promise<User> => {
-		const user = await this.getUser(userId);
-		if (!user) throw new Error();
 		const [updatedUser] = await this.database
 			.update(users)
 			.set({ level: sql`${users.level} + 1` })
-			.where(eq(users.userId, user.userId))
+			.where(eq(users.userId, userId))
 			.returning();
 		if (!updatedUser) throw new Error();
 		return updatedUser;
